pass ulong

+++ lib/ffi/gphoto2.rb
@@ -72,7 +72,7 @@ module FFI
     # int gp_file_get_data_and_size (CameraFile *file, const char **data, unsigned long int *size)
     attach_function :gp_file_get_data_and_size, [CameraFile.by_ref, :pointer, :pointer], :int
     # int gp_file_set_data_and_size (CameraFile *file, char *data, unsigned long int size)
-    attach_function :gp_file_set_data_and_size, [CameraFile.by_ref, :pointer, :int], :int
+    attach_function :gp_file_set_data_and_size, [CameraFile.by_ref, :pointer, :ulong], :int
 
     # gphoto2/gphoto2-list.h
     attach_function :gp_list_new, [:pointer], :int
